Revert the force_test changes
The original #21 wasn't working, and I tried to fix it via #23, but neither of these fully fixed the issue. I'm reverting the force flag until we have a chance to go through the original #2 and actually test that we're forcing testing correctly.

FWIW, my fix was also broken since `updated_templates` was an empty array, so even when forcing, we had nothing to test in our list.

This reverts commits:
  dfd1a05f0f55692dbb9dac9985d2f95cd97c032e
  9949115658ad2b5f9331ed15ea6c4bfae086e21d