Revert always generating test parent template
There is an expectation that AWS credentials and region are set
when actually running a test with Jetstream but not when executing
Jetstream locally to generate templates. The recent change to
always generate a master test template despite not initiating a test
breaks this behavior and jetstream will now error out if credentials and
a region are not set.