Update object comparison to ignore order
When comparing two CloudFormation templates we should only care if a
change is not ordering related.

For example:
obj1 = {"foo": "bar", "baz": "foo"}
obj2 = {"baz": "foo", "foo": "bar"}

When the above two objects are compared the should return as a match however
they are not a literal match when using the cmp() function due to the
order.